Remove dead code. Add Auction struct back
