+++ Junk4Dollars-app/Auction.swift
@@ -1,0 +1,6 @@
+struct Auction {
+    var title: String
+    var description: String
+    var starting_price: Int
+    var ends_at: String
+}
